# speed up electra attestation tests by 5x